requirements.txt
torch>=2.0.0
torchvision>=0.15.0
openmim>=0.3.9
mmengine>=0.10.1
mmcv>=2.0.0,<2.2.0
mmdet>=3.0.0
pyyaml>=6.0
pillow>=10.0.0
tqdm>=4.65.0
